requirement.txt
pytorch
torchaudio
pandas
numpy
matplotlib
librosa
spacy
transformers
kaggle